requirements.txt
fastapi[standard]
pydantic-settings
sqlalchemy
typing
bcrypt==4.3.0
pyjwt
uvicorn
